完整支持表约束

+++ GecodeTest/type_define.h
@@ -185,7 +185,7 @@ inline int func_eq(int *x) {
 }
 
 inline int func_abs(int *x) {
-	return fabsf(x[0]);
+	return (int)fabsf(x[0]);
 }
 
 inline int func_sub(int *x) {
